refactor(geo-layers): getTileLoadingState as TileLayer method
