+++ test/modules/geo-layers/get-tile-loading-state.spec.ts
@@ -4,38 +4,33 @@
 
 import {test, expect} from 'vitest';
 import {TileLayer} from '@deck.gl/geo-layers';
-import {getTileLoadingState} from '@deck.gl/geo-layers/tile-layer/get-tile-loading-state';
 
 const TEST_DATA = [
   {position: [0, 0]},
   {position: [1, 1]}
 ];
 
-test('getTileLoadingState#empty layer', () => {
+test('TileLayer#getTileLoadingState - empty layer', () => {
   const layer = new TileLayer({
     data: 'https://example.com/tiles/{z}/{x}/{y}',
     getTileData: () => null
   });
 
-  const state = getTileLoadingState(layer);
+  const state = layer.getTileLoadingState();
 
   expect(state.total).toBe(0);
   expect(state.loaded).toBe(0);
   expect(state.failed).toBe(0);
   expect(state.pending).toBe(0);
-  expect(state.percentLoaded).toBe(100);
-  expect(state.isComplete).toBe(true);
-  expect(state.isSuccess).toBe(true);
 });
 
-test('getTileLoadingState#all tiles loaded successfully', () => {
+test('TileLayer#getTileLoadingState - all tiles loaded successfully', () => {
   const layer = new TileLayer({
     id: 'test-layer',
     data: 'https://example.com/tiles/{z}/{x}/{y}',
     getTileData: async () => TEST_DATA
   });
 
-  // Mock tileset state with loaded tiles
   layer.state = {
     tileset: {
       selectedTiles: [
@@ -46,78 +41,67 @@ test('getTileLoadingState#all tiles loaded successfully', () => {
     }
   } as any;
 
-  const state = getTileLoadingState(layer);
+  const state = layer.getTileLoadingState();
 
   expect(state.total).toBe(3);
   expect(state.loaded).toBe(3);
   expect(state.failed).toBe(0);
   expect(state.pending).toBe(0);
-  expect(state.percentLoaded).toBe(100);
-  expect(state.isComplete).toBe(true);
-  expect(state.isSuccess).toBe(true);
 });
 
-test('getTileLoadingState#some tiles failed', () => {
+test('TileLayer#getTileLoadingState - some tiles failed', () => {
   const layer = new TileLayer({
     id: 'test-layer',
     data: 'https://example.com/tiles/{z}/{x}/{y}',
     getTileData: async () => TEST_DATA
   });
 
-  // Mock tileset with successful and failed tiles
   layer.state = {
     tileset: {
       selectedTiles: [
-        {isLoaded: true, content: TEST_DATA},   // Success
-        {isLoaded: true, content: null},         // Failed (404, error, etc.)
-        {isLoaded: true, content: TEST_DATA},   // Success
-        {isLoaded: true, content: null}          // Failed
+        {isLoaded: true, content: TEST_DATA},
+        {isLoaded: true, content: null},
+        {isLoaded: true, content: TEST_DATA},
+        {isLoaded: true, content: null}
       ]
     }
   } as any;
 
-  const state = getTileLoadingState(layer);
+  const state = layer.getTileLoadingState();
 
   expect(state.total).toBe(4);
   expect(state.loaded).toBe(2);
   expect(state.failed).toBe(2);
   expect(state.pending).toBe(0);
-  expect(state.percentLoaded).toBe(50);
-  expect(state.isComplete).toBe(true);
-  expect(state.isSuccess).toBe(false);
 });
 
-test('getTileLoadingState#tiles still loading', () => {
+test('TileLayer#getTileLoadingState - tiles still loading', () => {
   const layer = new TileLayer({
     id: 'test-layer',
     data: 'https://example.com/tiles/{z}/{x}/{y}',
     getTileData: async () => TEST_DATA
   });
 
-  // Mock tileset with loading tiles
   layer.state = {
     tileset: {
       selectedTiles: [
-        {isLoaded: true, content: TEST_DATA},   // Loaded
-        {isLoaded: false, content: null},        // Still loading
-        {isLoaded: false, content: null},        // Still loading
-        {isLoaded: true, content: null}          // Failed
+        {isLoaded: true, content: TEST_DATA},
+        {isLoaded: false, content: null},
+        {isLoaded: false, content: null},
+        {isLoaded: true, content: null}
       ]
     }
   } as any;
 
-  const state = getTileLoadingState(layer);
+  const state = layer.getTileLoadingState();
 
   expect(state.total).toBe(4);
   expect(state.loaded).toBe(1);
   expect(state.failed).toBe(1);
   expect(state.pending).toBe(2);
-  expect(state.percentLoaded).toBe(25);
-  expect(state.isComplete).toBe(false);
-  expect(state.isSuccess).toBe(false);
 });
 
-test('getTileLoadingState#all tiles failed', () => {
+test('TileLayer#getTileLoadingState - all tiles failed', () => {
   const layer = new TileLayer({
     id: 'test-layer',
     data: 'https://example.com/tiles/{z}/{x}/{y}',
@@ -126,24 +110,20 @@ test('getTileLoadingState#all tiles failed', () => {
     }
   });
 
-  // Mock tileset where all tiles failed
   layer.state = {
     tileset: {
       selectedTiles: [
-        {isLoaded: true, content: null},  // Failed
-        {isLoaded: true, content: null},  // Failed
-        {isLoaded: true, content: null}   // Failed
+        {isLoaded: true, content: null},
+        {isLoaded: true, content: null},
+        {isLoaded: true, content: null}
       ]
     }
   } as any;
 
-  const state = getTileLoadingState(layer);
+  const state = layer.getTileLoadingState();
 
   expect(state.total).toBe(3);
   expect(state.loaded).toBe(0);
   expect(state.failed).toBe(3);
   expect(state.pending).toBe(0);
-  expect(state.percentLoaded).toBe(0);
-  expect(state.isComplete).toBe(true);
-  expect(state.isSuccess).toBe(false);
 });
